Refactor project specific env hooks
There was a lot of duplication in the go and ruby workspace helpers, which I
think can probably be extracted into a utility that could be published. You
can now specify environment hooks to run when entering and exiting a directory
with a specified file, e.g.

    function enter_ruby_version {
      # do ruby stuff
    }

    function exit_ruby_version {
      # undo ruby_stuff
    }

    register_env_hook .ruby-version ruby_version

We define hooks for entering and exiting a directory, then register these
hooks to be run depending on the presence (or otherwise) of a .ruby-version
file.