Revert "Experiment: add ! and ? to ruby vim keywords"
This reverts commit 34b592b98268e1d0fa766714be7cef3e34a324ea.

I'd seen a PR for this in vim-ruby[1] that was rejected by tpope, saying:

    Adding ! to iskeyword isn't really accurate and more importantly has
    numerous problems in practice.

I didn't know what those problems were so I thought I'd find out. The
main one so far is that ! is a fairly common prefix operator, which
leads to problems:

    IS_ENABLED = true

    if !IS_ENABLED
       ^^^^^^^^^^^ all one keyword
    end

Presumably there are other problems but that's the first one I found.

[1] https://github.com/vim-ruby/vim-ruby/issues/341