Tidy up chruby and nvm initialisation
I had nvm init in a function for performance reasons I think, but it's
negligible now and doesn't seem to work properly inside the function for
reasons.

chruby init needs to know where homebrew is now, so I'm putting it in
the darwin bashrc since I don't have a linux box and can sort it out
if I ever get one.