Add git-add-repo script to import external repo histories
I'm doing some combining of repos at the moment, effectively mounting an
external repo at a subdirectory of the new host repo, and I want to
bring the git history along with that.

This pinches a script from SO[1] along with a few tweaks which does just
what I want, and it seems generally useful so let's keep it around.
which does just what I want, and it seems generally useful so let's keep
it around. which does just what I want, and it seems generally useful so
let's keep it around. which does just what I want, and it seems
generally useful so let's keep it around.

[1] https://stackoverflow.com/questions/1683531/how-to-import-existing-git-repository-into-another/43345686#43345686