Use tmux- and vim-native clipboard integrations
And bin reattach-to-user-namespace, hurrah (but thank you for your
service 🫡)

Relies on the terminal emulator having OSC52 support, which iTerm does.
External is actually the default, and means only tmux can access the
clipboard, not programs running under it. This is good because I don't
want untrusted programs seeing in my clipboard.

Vim is separately integrated via its own config.

https://github.com/tmux/tmux/wiki/Clipboard